pyproject.toml
[project]
name = "protein"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "dask>=2025.7.0",
    "ipykernel>=6.30.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "rcsb-api>=1.4.0",
    "swifter==1.4.0",
]
